refactor: move internal class to internal folder
