+++ packages/core/src/animation/Animator.ts
@@ -4,16 +4,16 @@ import { Component } from "../Component";
 import { Entity } from "../Entity";
 import { ClassPool } from "../RenderPipeline/ClassPool";
 import { Transform } from "../Transform";
-import { AnimationCureOwner } from "./AnimationCureOwner";
+import { AnimationCureOwner } from "./internal/AnimationCureOwner";
 import { AnimatorController } from "./AnimatorController";
 import { AnimatorControllerLayer } from "./AnimatorControllerLayer";
-import { AnimatorLayerData } from "./AnimatorLayerData";
-import { AnimatorStateData } from "./AnimatorStataData";
+import { AnimatorLayerData } from "./internal/AnimatorLayerData";
+import { AnimatorStateData } from "./internal/AnimatorStataData";
 import { AnimatorState } from "./AnimatorState";
-import { AnimatorStatePlayData } from "./AnimatorStatePlayData";
+import { AnimatorStatePlayData } from "./internal/AnimatorStatePlayData";
 import { AnimatorStateTransition } from "./AnimatorTransition";
 import { AnimatorUtils } from "./AnimatorUtils";
-import { CrossCurveData } from "./CrossCurveData";
+import { CrossCurveData } from "./internal/CrossCurveData";
 import { AnimationProperty } from "./enums/AnimationProperty";
 import { AnimatorLayerBlendingMode } from "./enums/AnimatorLayerBlendingMode";
 import { InterpolableValueType } from "./enums/InterpolableValueType";
