+++ packages/core/src/animation/internal/AnimatorStataData.ts
@@ -1,4 +1,4 @@
-import { Component } from "../Component";
+import { Component } from "../../Component";
 import { AnimationCureOwner } from "./AnimationCureOwner";
 
 /**
