+++ packages/core/src/animation/internal/CrossCurveData.ts
@@ -1,4 +1,4 @@
-import { Component } from "../Component";
+import { Component } from "../../Component";
 import { AnimationCureOwner } from "./AnimationCureOwner";
 /**
  * @internal
